// src/types/plant.ts
export type Plant = {
    index: number;
    family: string;
    categories: string;
    origin: string;
    climate: string;
    img_url: string;
    name: string;
    water_category: string;
    venomous: boolean;
    size: string;
    soil: string;
    sunlight: string;
    experience_level: number;
    group_name: string;
    compatibility: number;
    distance: number;
};